Bajaj Hackathon/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
PyPDF2==3.0.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
pdfminer.six==20221105
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
